points counter

+++ server/src/main/java/com/mentalmath/server/service/GameService.java
@@ -2,5 +2,5 @@ package com.mentalmath.server.service;
 
 public interface GameService {
     String generateQuestion();
-    String checkAnswer(int answer);
+    String checkAnswer(double answer);
 }
